Add documentation and rename folder for consistency
