+++ doc/user_guide/examples/orm/create_tables.rst
@@ -1,0 +1,8 @@
+.. _example_orm_create_table:
+
+Creating Tables
+===============
+
+.. literalinclude:: ../../../../examples/features/orm/_0_create_tables.py
+       :language: python3
+       :caption: examples/features/orm/_0_create_tables.py
