+++ doc/user_guide/examples/orm/index.rst
@@ -1,0 +1,10 @@
+.. _examples_orm:
+
+ORM Examples
+============
+
+.. toctree::
+   :maxdepth: 2
+
+   create_tables
+   insert_data
