+++ doc/user_guide/examples/orm/insert_data.rst
@@ -1,0 +1,8 @@
+.. _example_orm_insert_data:
+
+Inserting Data
+==============
+
+.. literalinclude:: ../../../../examples/features/orm/_1_add_example_data.py
+       :language: python3
+       :caption: examples/features/orm/_1_add_example_data.py
